+++ src/interface/settings/general_settings.py
@@ -1,9 +1,12 @@
 # Interface/Settings/general_settings.py
 from PySide6.QtWidgets import QMessageBox
+from tkinter import messagebox
 import shutil
 import customtkinter as ctk
 import src.utils.fonts as fonts
 from CTkToolTip import CTkToolTip
+import subprocess
+import sys
 from src.utils.save_settings import save_all_settings
 from src.utils.load_settings import load_data_path
 from src.utils.toast import show_toast
@@ -122,11 +125,15 @@ def create_general_settings_tab(globals, general_frame):
                     values=theme_labels,
                     state="readonly",
                     width=150).pack(side="left")
+    
+    ctk.CTkLabel(theme_frame,
+                 text="*Requires restart",
+                 font=fonts.body_font).pack(side="left", padx=6, pady=0)
 
     # Version Check
     version_frame = ctk.CTkFrame(general_frame,
-                                  bg_color="transparent",
-                                  fg_color="transparent")
+                                 bg_color="transparent",
+                                 fg_color="transparent")
     version_frame.pack(fill="x", pady=10, padx=10)
 
     ctk.CTkLabel(version_frame,
@@ -148,7 +155,45 @@ def create_general_settings_tab(globals, general_frame):
                     variable=globals.github_check_var,
                     onvalue=True,
                     text=None,
-                    offvalue=False).pack(side="left", padx=5)
+                    width=0,
+                    offvalue=False).pack(side="left")
+    
+    ctk.CTkLabel(version_frame,
+                 text="*Requires restart",
+                 font=fonts.body_font).pack(side="left", padx=6, pady=0)
+    
+    # Beta Version
+    beta_frame = ctk.CTkFrame(general_frame,
+                              bg_color="transparent",
+                              fg_color="transparent")
+    beta_frame.pack(fill="x", pady=10, padx=10)
+
+    ctk.CTkLabel(beta_frame,
+                 text=None,
+                 image=None,
+                 width=40).pack(side="left", padx=6, pady=0)
+
+    beta_check_label = ctk.CTkLabel(beta_frame,
+                                    text="Include Beta Updates",
+                                    font=fonts.heading_font)
+    beta_check_label.pack(side="left", padx=(0, 12))
+    CTkToolTip(beta_check_label,
+               message="Inclides beta releases in version check",
+               delay=0.8,
+               follow=True,
+               padx=10,
+               pady=5)
+
+    ctk.CTkCheckBox(beta_frame,
+                    variable=globals.beta_var,
+                    onvalue=True,
+                    text=None,
+                    width=0,
+                    offvalue=False).pack(side="left")
+    
+    ctk.CTkLabel(beta_frame,
+                 text="*Requires restart",
+                 font=fonts.body_font).pack(side="left", padx=6, pady=0)
 
     # Context Detection
     context_frame = ctk.CTkFrame(general_frame,
@@ -212,15 +257,24 @@ def create_general_settings_tab(globals, general_frame):
 
     def delete_chats():
         # Create a messagebox asking for confirmation
-        reply = QMessageBox.question(
-            None,
-            "Delete Chats",
-            f"Are you sure you would like to delete all conversations?",
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-            QMessageBox.StandardButton.Yes)
-        if reply == QMessageBox.StandardButton.Yes:
-            shutil.rmtree(load_data_path("local", "chats"))
-            show_toast(globals, "Conversations deleted!")
+        if globals.qt_mode:
+            reply = QMessageBox.question(
+                None,
+                "Delete Chats",
+                f"Are you sure you would like to delete all conversations?",
+                QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+                QMessageBox.StandardButton.Yes)
+            if reply == QMessageBox.StandardButton.Yes:
+                shutil.rmtree(load_data_path("local", "chats"))
+                show_toast(globals, "Conversations deleted!")
+        else:
+            reply = messagebox.askyesno(
+                globals.root,
+                title="Delete Chats?",
+                message="Would you like to delete all saved chats?")
+            if reply:
+                shutil.rmtree(load_data_path("local", "chats"))
+                show_toast(globals, "Conversations deleted!")
 
     ctk.CTkLabel(deletion_frame,
                  text=None,
@@ -245,4 +299,35 @@ def create_general_settings_tab(globals, general_frame):
 
     ctk.CTkButton(save_button_frame,
                   text="Save Settings",
-                  command=lambda: save_all_settings(globals)).pack()
+                  command=lambda: save_button(globals)).pack()
+
+    def save_button(globals):
+        """Saves and prompts for restart if required."""
+        prompt_restart = False
+        if globals.github_check != globals.github_check_var.get():
+            prompt_restart = True
+        elif globals.active_theme != globals.theme_var.get():
+            prompt_restart = True
+        elif globals.beta != globals.beta_var.get():
+            prompt_restart = True
+        save_all_settings(globals)
+
+        if prompt_restart:
+            if globals.qt_mode:
+                reply = QMessageBox.question(
+                    None,
+                    "Restart Pearl?",
+                    f"Would you like to restart Pearl to apply all changes?",
+                    QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+                    QMessageBox.StandardButton.Yes)
+                if reply == QMessageBox.StandardButton.Yes:
+                    subprocess.Popen(globals.app_path)
+                    sys.exit(0)
+            else:
+                    reply = messagebox.askyesno(
+                        parent=globals.root,
+                        title="Restart Pearl",
+                        message="Would you like restart Pearl to apply all changes?")
+                    if reply:
+                        subprocess.Popen(globals.app_path)
+                        sys.exit(0)